requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
discord.py==2.3.2